+++ src/main/java/top/liheji/server/pojo/AuthGroupPermissions.java
@@ -23,19 +23,19 @@ public class AuthGroupPermissions implements Serializable {
      * 数据库字段
      */
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 外键
      */
-    private Integer groupId;
+    private Long groupId;
 
-    private Integer permissionId;
+    private Long permissionId;
 
     /**
      * 非数据库字段
      */
-    public AuthGroupPermissions(Integer groupId, Integer permissionId) {
+    public AuthGroupPermissions(Long groupId, Long permissionId) {
         this.groupId = groupId;
         this.permissionId = permissionId;
     }
